% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientDataBaseDetail.R
\name{orientDataBaseDetail}
\alias{orientDataBaseDetail}
\title{get details of an OrientDB DataBase}
\usage{
orientDataBaseDetail(orient, database, ...)
}
\arguments{
\item{orient}{An element created with \code{orientConnect}.}
}
\description{
\code{orientDataBaseDetail} Execute a SQL query against the OrientDB database.
}
\author{
Mahmood S. Zargar \email{mahmood@gmail.com}
Mohamed Karim Bouaziz \email{mohamed.karim.bouaziz@gmail.com}
}
\seealso{
\code{\link{orientDataBases}}, \code{\link{orientConnect}}
}

