% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientDataBases.R
\name{orientDataBases}
\alias{orientDataBases}
\title{Extract names of all OrientDB DataBases.}
\usage{
orientDataBases(orient)
}
\arguments{
\item{orient}{An element created with \code{orientConnect}.}
}
\description{
\code{orientDataBases} Extract names of all OrientDB DataBases.
}
\author{
Mahmood S. Zargar \email{mahmood@gmail.com}
Mohamed Karim Bouaziz \email{mohamed.karim.bouaziz@gmail.com}
}
\seealso{
\code{\link{orientConnect}}, \code{\link{orientDataBaseDetail}}
}

