% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientGetDocumentByClass.R
\name{orientGetDocumentByClass}
\alias{orientGetDocumentByClass}
\title{Retive a document from orientdb}
\usage{
orientGetDocumentByClass(orient, database, className, recordPosition, ...)
}
\arguments{
\item{orient}{An element created with \code{orientConnect}.}

\item{database}{The database to be used for retrieving a document.}

\item{recordPosition}{recordPosition of your orient node e.g. a node has the id #12:5 where 12 is your classid
and 5 is your recordPosition}
}
\description{
\code{orientGetDocumentByClass} retrieves a document form orient
}
\author{
Mohamed Karim Bouaziz \email{mohamed.karim.bouaziz@gmail.com}
}
\seealso{
\code{\link{orientConnect}}, \code{\link{orientDataBases}}, \code{\link{orientDataBaseDetail}}
}

