% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientSql.R
\name{orientSql}
\alias{orientSql}
\title{Execute a SQL query against the OrientDB database}
\usage{
orientSql(orient, database, query, batch = -1, conv.dates = TRUE,
  date.fmt = "ymd", auto.na = TRUE, rm.meta = TRUE, conv.rid = FALSE,
  unwind = FALSE, formats = c(), verbos = FALSE, ...)
}
\arguments{
\item{orient}{An element created with \code{orientConnect}.}

\item{database}{The database to be user in SQL Query.}

\item{batch}{The number of record in return data (in SELECT Query), default is \code{-1} it means unlimited records.}

\item{conv.dates}{}

\item{formats}{}
}
\description{
\code{orientSql} Execute a SQL query against the OrientDB database.
}
\author{
Mahmood S. Zargar \email{mahmood@gmail.com}
Mohamed Karim Bouaziz \email{mohamed.karim.bouaziz@gmail.com}
}
\seealso{
\code{\link{orientConnect}}, \code{\link{orientDataBases}}, \code{\link{orientDataBaseDetail}}
}

