% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientConnect.R
\name{orientConnect}
\alias{orientConnect}
\title{Connect to OrientDB DataBase}
\usage{
orientConnect(host = "localhost", username = "admin", password = "admin",
  port = "2480")
}
\arguments{
\item{host}{The host name to the OrientDB DataBase (by default "localhost").}

\item{username}{The username of the OrientDB DataBase.}

\item{password}{The password of the OrientDB DataBase.}

\item{port}{The Port of OrientDB DataBase (by default 2480)}
}
\description{
\code{orientConnect} Create an OrientDB connection string to be used in further
 commands
}
\author{
Mohamed Karim Bouaziz \email{mohamed.karim.bouaziz@gmail.com}
}
\seealso{
\code{\link{orientDataBases}}, \code{\link{orientDataBaseDetail}}
}

